# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "ActionKind": {
        "description": "What the planner can decide to do.",
        "enum": ["reorder", "dispatch", "notify", "escalate"],
        "title": "ActionKind",
        "type": "string"
      },
      "ActionOut": {
        "description": "A planner decision and its review state.",
        "properties": {
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "estimated_cost_kobo": {
            "title": "Estimated Cost Kobo",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/ActionKind"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "requires_review": {
            "title": "Requires Review",
            "type": "boolean"
          },
          "review_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Note"
          },
          "review_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Review Reason"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed At"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          },
          "run_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "$ref": "#/components/schemas/ActionStatus"
          },
          "store_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Store Id"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "run_id",
          "store_id",
          "kind",
          "status",
          "requires_review",
          "review_reason",
          "payload",
          "estimated_cost_kobo",
          "confidence",
          "rationale",
          "reviewed_by",
          "reviewed_at",
          "review_note",
          "created_at",
          "updated_at"
        ],
        "title": "ActionOut",
        "type": "object"
      },
      "ActionStatus": {
        "description": "Lifecycle of a planner decision.",
        "enum": ["proposed", "pending_review", "approved", "rejected", "executed"],
        "title": "ActionStatus",
        "type": "string"
      },
...
      "Body_upload_photo_v1_shelves__shelf_id__photos_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "description": "JPEG, PNG or WebP shelf photo",
            "title": "File",
            "type": "string"
          }
        },
        "required": ["file"],
        "title": "Body_upload_photo_v1_shelves__shelf_id__photos_post",
        "type": "object"
      },
...
      "DetectedItem": {
        "description": "One product group seen on the shelf.",
        "properties": {
          "confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "facings": {
            "description": "Front-facing units visible.",
            "minimum": 0.0,
            "title": "Facings",
            "type": "integer"
          },
          "label": {
            "description": "What is visible, e.g. 'Peak Evaporated Milk 160g'.",
            "title": "Label",
            "type": "string"
          },
          "region": {
            "$ref": "#/components/schemas/BBox"
          },
          "sku_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Planogram SKU id this product matches, or null if not in the planogram.",
            "title": "Sku Id"
          }
        },
        "required": ["sku_id", "label", "facings", "region", "confidence"],
        "title": "DetectedItem",
        "type": "object"
      },
      "ErrorResponse": {
        "description": "Body of every 4xx/5xx the API raises deliberately.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": ["detail"],
        "title": "ErrorResponse",
        "type": "object"
      },
      "ExtractionOut": {
        "description": "A finished vision run: the model's extraction, the derived summary, and its cost.",
        "properties": {
          "attempts": {
            "title": "Attempts",
            "type": "integer"
          },
          "cost_usd": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost Usd"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "extraction": {
            "$ref": "#/components/schemas/ShelfExtraction"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "input_tokens": {
            "title": "Input Tokens",
            "type": "integer"
          },
          "latency_ms": {
            "title": "Latency Ms",
            "type": "integer"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "output_tokens": {
            "title": "Output Tokens",
            "type": "integer"
          },
          "overall_confidence": {
            "title": "Overall Confidence",
            "type": "number"
          },
          "planogram_version": {
            "title": "Planogram Version",
            "type": "integer"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          },
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "summary": {
            "$ref": "#/components/schemas/ExtractionSummary"
          }
        },
        "required": [
          "id",
          "run_id",
          "model",
          "provider",
          "planogram_version",
          "overall_confidence",
          "extraction",
          "summary",
          "input_tokens",
          "output_tokens",
          "cost_usd",
          "latency_ms",
          "attempts",
          "created_at"
        ],
        "title": "ExtractionOut",
        "type": "object"
      },
      "ExtractionSummary": {
        "description": "Numbers the dashboard and planner consume; derived, never model-authored.",
        "properties": {
          "compliance_rate": {
            "description": "Share of planogram slots at or above their minimum facings.",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Compliance Rate",
            "type": "number"
          },
          "planogram_share_of_shelf": {
            "description": "Percent of visible facings that belong to planogram SKUs.",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Planogram Share Of Shelf",
            "type": "number"
          },
          "slots": {
            "items": {
              "$ref": "#/components/schemas/SlotCompliance"
            },
            "title": "Slots",
            "type": "array"
          },
          "stock_out_count": {
            "title": "Stock Out Count",
            "type": "integer"
          },
          "stock_out_rate": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Stock Out Rate",
            "type": "number"
          },
          "unknown_item_count": {
            "title": "Unknown Item Count",
            "type": "integer"
          }
        },
        "required": [
          "slots",
          "stock_out_count",
          "stock_out_rate",
          "compliance_rate",
          "planogram_share_of_shelf",
          "unknown_item_count"
        ],
        "title": "ExtractionSummary",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PhotoOut": {
        "description": "A shelf photo and where its processing stands.",
        "properties": {
          "content_type": {
            "title": "Content Type",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "description": "Time-limited URL for the original image.",
            "title": "Download Url",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "extraction": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExtractionOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "shelf_id": {
            "format": "uuid",
            "title": "Shelf Id",
            "type": "string"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "status": {
            "$ref": "#/components/schemas/PhotoStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "uploaded_by": {
            "title": "Uploaded By",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "shelf_id",
          "status",
          "content_type",
          "size_bytes",
          "width",
          "height",
          "uploaded_by",
          "error",
          "created_at",
          "updated_at",
          "download_url",
          "extraction"
        ],
        "title": "PhotoOut",
        "type": "object"
      },
      "PhotoStatus": {
        "description": "Where a photo is in the pipeline.",
        "enum": ["queued", "processing", "done", "failed"],
        "title": "PhotoStatus",
        "type": "string"
      },
...
      "ShareOfShelf": {
        "description": "Share of visible facings held by one planogram SKU.",
        "properties": {
          "percent": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percent",
            "type": "number"
          },
          "sku_id": {
            "format": "uuid",
            "title": "Sku Id",
            "type": "string"
          }
        },
        "required": ["sku_id", "percent"],
        "title": "ShareOfShelf",
        "type": "object"
      },
      "ShelfExtraction": {
        "description": "What the vision agent returns. The contract for evals and the review queue.",
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/DetectedItem"
            },
            "title": "Items",
            "type": "array"
          },
          "notes": {
            "description": "Anything a reviewer should know: glare, occlusion, doubts.",
            "title": "Notes",
            "type": "string"
          },
          "overall_confidence": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Overall Confidence",
            "type": "number"
          },
          "share_of_shelf": {
            "items": {
              "$ref": "#/components/schemas/ShareOfShelf"
            },
            "title": "Share Of Shelf",
            "type": "array"
          },
          "stock_outs": {
            "description": "Planogram SKU ids with zero facings visible.",
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Stock Outs",
            "type": "array"
          }
        },
        "required": ["items", "stock_outs", "share_of_shelf", "overall_confidence", "notes"],
        "title": "ShelfExtraction",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": ["loc", "msg", "type"],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Shelf photos and cold-chain telemetry in, reviewed decisions out. The contract is written first in apps/api/openapi.yaml; this document is generated from the implementation and checked against it in CI.",
    "title": "ShelfSense API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/actions": {
      "get": {
        "description": "Actions, newest first, filterable by status, kind and store.",
        "operationId": "list_actions_v1_actions_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ActionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ActionKind"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "store_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Store Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ActionOut"
                  },
                  "title": "Response List Actions V1 Actions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Actions",
        "tags": ["actions"]
      }
    },
    "/v1/actions/{action_id}": {
      "get": {
        "description": "One action.",
        "operationId": "read_action_v1_actions__action_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "action_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Action Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ActionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Action",
        "tags": ["actions"]
      }
    },
...
    "/v1/photos/{photo_id}": {
      "get": {
        "description": "One photo with its extraction once processing is done.",
        "operationId": "read_photo_v1_photos__photo_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "photo_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Photo Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read Photo",
        "tags": ["photos"]
      }
    },
...
    "/v1/shelves/{shelf_id}/photos": {
      "get": {
        "description": "Photos of a shelf, newest first.",
        "operationId": "list_photos_v1_shelves__shelf_id__photos_get",
        "parameters": [
          {
            "in": "path",
            "name": "shelf_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shelf Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PhotoOut"
                  },
                  "title": "Response List Photos V1 Shelves  Shelf Id  Photos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Photos",
        "tags": ["photos"]
      },
      "post": {
        "description": "Store the photo and queue the vision extraction. Returns 202 with status ``queued``.",
        "operationId": "upload_photo_v1_shelves__shelf_id__photos_post",
        "parameters": [
          {
            "in": "path",
            "name": "shelf_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Shelf Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_v1_shelves__shelf_id__photos_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhotoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not allowed"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Already exists"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Too large"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not an image"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Photo",
        "tags": ["photos"]
      }
    },
...
  },
...
}